perf(coordinate): 静的ヘッダを即時描画しデータ部分をSuspenseでストリーミング
/style と同様、/coordinate の遷移時に静的タイトル/説明までスケルトンで
覆われていた問題を解消。先頭の await(connection / 認証 / プロフィール)を
やめ、静的ヘッダを即描画。ユーザー依存部分を分離してストリーミングする。

- CoordinatePageBody: 認証・残高・生成フォーム・生成結果・RefreshOnMount を
  隔離し <Suspense> 配下に。connection() もこの動的領域内へ移動
- 本番ビルドで低速回線でもタイトル+説明が先に実テキスト表示され、本文だけ
  スケルトン→差し替えを確認
